refactor: Implemented modularity to work with SSR
