+++ app/(dashboard)/problem-statements/components/CountdownWrapper.jsx
@@ -1,0 +1,8 @@
+"use client";
+
+import CountdownTimer from "@/app/components/CountdownTimer";
+
+// Simple client component wrapper for CountdownTimer
+export default function CountdownWrapper() {
+    return <CountdownTimer />;
+}
