+++ app/(dashboard)/rules/data/data.js
@@ -1,0 +1,14 @@
+// Rules set
+const rules = [
+    "All team members must be registered on the Kodesphere platform before the start of the hackathon.",
+    "Code submissions must be original work created during the hackathon period.",
+    "Teams may use open-source libraries and frameworks, but must disclose all dependencies.",
+    "Submissions must address the provided problem statement and meet the minimum requirements.",
+    "All code must be submitted before the deadline. Late submissions will not be accepted."
+];
+
+// Export function to get rules
+export function getRules() {
+    return rules;
+}
+
