+++ app/components/client/ClientCountdownTimer.jsx
@@ -1,0 +1,7 @@
+"use client";
+
+import CountdownTimer from "@/app/components/CountdownTimer";
+
+export default function ClientCountdownTimer({ targetDate }) {
+    return <CountdownTimer targetDate={targetDate} />;
+}
